fix: run the Unity auto-API-updater for Unity 2018.3.0f2
In an attempt to upgrade the code to Unity version 2018.3.0f2.

The auto-API-update wasn't entirely successful, in the sense that there are
still build warnings and errors that appear to be caused by the use of a newer
Unity version than the creator of the sketchfab/UnityGLTF repo. I believe the
creator was using some version of Unity 2017, although I'm not sure about that.