update: Assets/Sketchfab-UnityGLTF from upstream git repo
Update the files under Assets/Sketchfab-UnityGLTF from the the master branch of
the upstream git repo at https://github.com/sketchfab/UnityGLTF.  The HEAD
commit on the master branch was `c54fd454859c9ef8e1244c8d08c3f90089768702`
(timestamp: Wed Jun 26 11:03:38 2019 +0200).

Note that my `Assets/Sketchfab-UnityGLTF` maps to the `UnityGLTF` subdirectory
of the sketchfab/UnityGLTF repo. I am intentionally excluding the
`GLTFSerialization` subdirectory of the `sketchfab/UnityGLTF` repo, which is a
mirror of Khronos' UnityGLTF project for building `GLTFSerialization.dll`.
Including the `GLTFSerialization` folder in a Unity project causes build errors
and must be instead be built from the command line using the `dotnet` tool.

I wanted to update the Sketchfab source files because I am about to attempt a
refactoring of the `GLTFEditorImporter` class in order to remove its
dependencies on `UnityEditor`. I wanted to update to the latest version of the
source files before doing that, so that I have the option of submitting my
changes back to Sketchfab as a pull request against
https://github.com/sketchfab/UnityGLTF.

The source file changes that have occurred since I originally imported the files
is pretty trivial.

I updated the files by cloning/pulling the latest version of
https://github.com/sketchfab/UnityGLTF in a separate directory and then running
the following rsync command from the root of my git repo:

rsync --recursive --delete --checksum --verbose --exclude='*.meta' \
  --exclude='Plugins/' --exclude='Library/' --exclude='UnityPackageManager/' \
  /mnt/c/Users/Ben/git/sketchfab/UnityGLTF/UnityGLTF/ \
  Assets/Sketchfab-UnityGLTF/

Note to self: Remember to run the rsync command with the `--dry-run` option
first, before running it for real!