add: GLTFRuntimeImporterCache.cs
I added this as a replacement for `AssetCache` in `GLTFEditorImporter` /
`GLTFRuntimeImporterCache`.

The purpose of `AssetCache` / `GLTFRuntimeImporterCache` is to store Unity
objects (e.g. textures, materials, meshes) in memory when they are loaded from
the GLTF file, until such a time that they can be hooked together into the final
GameObject hierarchy for the overall scene/model.

I created `GLTFRuntimeImporterCache` because the data structures used in
`AssetCache` were more complex than needed, and were causing unnecessary
obstacles when refactoring the `GLTFEditorImporter` / `GLTFRuntimeImporter`
code.