add: GLTFImporter (new parent class of runtime & editor importers)
Formerly, `GLTFRuntimeImporter` was the parent class of `GLTFEditorImporter`,
but now I am introducing `GLTFImporter` as a common parent class of
`GLTFRuntimeImporter` and `GLTFEditorImporter`.  Most of the importing code
now lives in `GLTFImporter`.

I wanted to make `GLTFRuntimeImporter` a child of `GLTFImporter` so that I could
specialize the constructor of `GLTFRuntimeImporter` and add the build-time
shader configuration as a parameter. This will enable me to check what shaders
variants/keywords are available at runtime and to catch errors while I'm
importing materials.