GLTFRuntimeImporter: use shader config to select shader variant
Add `GLTFRuntimeShaderConfig` as a require constructor parameter.
`GLTFRuntimeShaderConfig` is used to construct the list of Standard shader
variants to include in the build, and to query the list of available shader
variants at runtime.

When importing a material, use `GLTFRuntimeShaderConfig` (i.e. `_shaderConfig`)
to find and load a shader variant whose keywords are a minimal superset of the
keywords enabled on the material.