checkpoint: correctly identify clicked item in Project Browser
Do this by building a mapping of asset guids -> screen rects on
each frame.

This commit is a checkpoint towards the goal of being able to drag GLTF files
into the right pane of the Project Browser when the right pane is empty (i.e. an
empty directory is selected in the folder tree on the left).