GameManager: remove redundant StartImport call on startup [cleanup]
An import job to load the default Piglet model is now started from
inside ParseCommandLineArgs, provided that no other glTF model has
been specified with the --load option. As such, the import job that is
started in Start() is no longer needed. (It is immediately replaced
by the import job that is started inside ParseCommandLineArgs.)