Singleton.cs: fix NullReferenceException in Editor scripts [bugfix]
I'm not sure how, but sometimes m_ShuttingDown gets set to true when
the Singleton class is being using from an Editor script, which causes
null to be returned from the Singleton.Instance method.

This commit fixes the problem by simply ignoring the m_ShuttingDown
when running inside the Editor.