remove: WebGlTextureImporter [cull]
Remove the specialized texture importer for
WebGL (WebGlTextureImporter), which creates/loads textures on the
Javascript/WebGL side instead of using UnityWebRequestTexture as done
on desktop/Android. So far, it seems that WebGlTextureImporter does
not decrease the overall import time very much (if at all), while it
adds a lot of potential for errors/bugs.

One existing issue with WebGlTextureImporter is that it somehow causes
hangs and/or NullReferenceExceptions when importing .zip files, and I
don't understand why.  I have also intermittently seen the wrong image
data being shown/loaded for certain textures (and I don't understand
that either). I suspect the latter issue has something to do with
loading the textures in parallel, but I may need to gain more
experience with OpenGL/WebGL programming before I can understand why
it is happening.

That said, WebGlTextureImporter is a good start and a potential source
of future performance improvements. I may resurrect it at a later
time.