WebGL: fix error SendMessage calling the wrong method [bugfix]
In the WebGL build, SendMessage was throwing an error saying that
GameManager.StartImport was being called with the wrong number/type of
arguments. The problem was that the StartImport method is overloaded
and SendMessage doesn't correctly handle method
overloading. (SendMessage simply invokes the first method that matches
the given method name).  I fixed the issue by creating a new wrapper
method around StartImport (ImportUrlWebGl) that isn't overloaded.