Android: prevent model reload when switching back to app [bugfix]
This commit fixes a problem where the Android app would reload the
current model every time the user brought the app from the background
to the foreground.

The solution implemented in this commit is to remember the Andoid
intent URI from the previous time the app had focus, and compare it
against the current intent URI to see if it has changed.