move Assets/Plugins files into Piglet/Plugins and PigletViewer/Plugins [refactor]
Where possible, move files under Assets/Plugins into
Piglet/Plugins or PigletViewer/Plugins.

The motivation for this reorganization was the fact that the Unity
Asset Store tools only allows selecting/uploading a single folder for
submission to the asset store.

It has the side benefit that it makes the project tidier.