add separate WebGL template (index.html) for Unity 2020+ [new]
My original index.html works for Unity 2018 and Unity 2019 but does
not work in Unity 2020 because there have been substantial changes to
the WebGL pipeline and JavaScript API. For a detailed description of
the WebGL-related changes introduced in Unity 2020, see:

https://forum.unity.com/threads/changes-to-the-webgl-loader-and-templates-introduced-in-unity-2020-1.817698/

Initially I tried to upload a separate package to the Unity Asset
Store for Unity 2020+, but this did not work because the Asset Store
Tools are current broken in Unity 2020. I decided to work around this
issues by bundling two WebGL templates ("Piglet2018" and "Piglet2020")
and adding an explanation to the manual. (I will add the update to
the manual PDF in a future commit.)