replace "Piglet" title mesh/camera with IMGUI texture [cleanup]
The Piglet title mesh was used to render "Piglet" in fancy writing
at the top of UI window.  I used a dual-camera setup with layers
to ensure that the title mesh was drawn on top of everything else.

For some reason, this setup does not work properly when the active
render pipeline is changed to URP. (The title camera just draws
over the entire screen so that the UI and model become invisible.)
To fix this, I replaced the title mesh with a PNG image that is
drawn directly by the IMGUI code. This setup is a lot simpler and
the resulting title graphic looks better too.