Gui: fix play/pause functionality [bugfix]
In the Editor, the time of the selected animation clip would continue
to advance in the background while the clip was paused.

I fixed this by setting the playback speed to zero to pause the
clip (clip.speed = 0), rather than disabling the Animation
component (anim.enabled = false).

These changes also fix a problem where the model would not be
correctly reset to its default pose when selecting "Static Pose".