Gui: reset to static pose between animation clips [bugfix]
We need to reset to the static pose before switching to a new
animation clip because the animation clips may (or may not) rely on
starting from the static pose.

The only time that animation clip can be correctly started from any
pose is when it exhaustively specifies all position/rotation/scale
values for all game objects in the hierarchy.