GameManager: use GltfImportOptions to automatically scale models [new]
Use the new GltfImportOptions parameter of GetImportTask to
automatically scale the imported models to a standard size.

Previously the resizing of the imported models was done in
PigletViewer (in ModelManager), but I decided it would be better to
move that code into Piglet and make it available through an option.