CameraBehavior: position/orient camera relative to model [refactor][bugfix]
When first loading a model, reset the camera position and orientation
relative to the model, rather than the other way around.

This fixes a bug where the model is positioned at the wrong distance
from the camera when resetting a model to its "Static Pose" after
playing an animation clip.

It also makes the code simpler and easier to understand in general.