ModelManager: fix SpinModel performance [perf]
SpinModel was hurting performance because it was recalculating the
world-space axis-aligned bounding box of the model in every frame, by
traversial the entire hierachy of the model.

It was doing this to determine the center point of the model, to use
as a pivot point for rotations.

I fixed the problem by calculating the bounding box and center
point (`_modelCenterPoint`) only once, when the model is first loaded.