ProjectSettings: upgrade project to Unity 2018.4.20f1 [upgrade]
Upgrade project version from Unity 2018.4.16f1 -> 2018.4.20f1.

I upgraded because there are changes to the drag-and-drop code in
Piglet 1.3.3 that cause Unity 2018.4.16f1 to hang.

I was able to consistently reproduce the problem by
dragging-and-dropping the "cartoon_hartman" folder into the Editor
twice. The Editor hangs during the second drag-and-drop import. I'm
not exactly why this happens but I suspect it has something to do with
deleting files and the =AssetDatabase.Refresh= loop.

I did not do an exhaustive binary search on Unity versions, so I don't
know whether Unity versions 2018.4.17f1, 2018.4.18f1, 2018.4.19f1 are
also subject to this issue.