add URP RenderPipelineAsset and Renderer [new][devops]
Toward the goal of automatically building/testing/profiling Piglet
with URP, add the RenderPipelineAsset and RenderPipelineAsset_Renderer
asset files to the project, so that I do not have to create them in an
Editor script.

Happily, adding these files in the Unity project does not cause any
compile errors if the URP
package (com.unity.render-pipelines.universal) is not installed. It
also does not cause any compile errors if the current Unity version
predates URP (e.g. Unity 2018.4).

I tested these conditions using Unity 2018.4.20f1 and Unity
2019.3.13f1 without the URP package installed. (The Unity project in
this repo uses Unity 2018.4.20f1 and this is also the minimum Unity
version that is supported by Piglet.)