command line: replace --prompt with --button [refactor]
My original implementation of --prompt wasn't very useful because the
program would continue after any mouse click, including when I tried
to rotate the model with the left-click and drag.

Using a button allows me to inspect each model as much as I like
before continuing to the next one.

Note: Pressing Return or Space also causes the program to continue, in
addition to clicking the button.