Gui.cs: switch to automatic positioning of GUI elements [WIP]
Switch from manual position of GUI text/controls to automatic
positioning using GUILayout.

I'm making these changes so that in future commits, I will be able to
easily set custom font sizes and screen margins on iOS, without
breaking the layout of the GUI elements. Custom font sizes are
desirable on iOS because the screen DPI is generally higher than on
Android. Custom screen margins are desirable because the panel
containing the microphone and front-facing camera obscures the top
part of the screen. In addition, iOS apps overlay a handle widget
along the bottom part of the screen, used for closing the app, which
obscures the bottom part of the screen.

This is a [WIP] commit because the animation controls the following
issues still need to be fixed:

(1) the animation controls are currently in the wrong horizontal order
(2) the animation timeline slider does not expand to the full width
left over by the Play/Pause button and the combo box for selecting clips
(3) the margin at the bottom of the screen is currently too large