Gui.cs: scale fonts and GUI controls based on screen DPI [UX]
Scale the size of fonts and GUI controls based on the current
platform (e.g. PC, iOS) and screen DPI.

The main motivation for this change was to make the text larger and
more readable on iOS, but it should also make PigletViewer more usable
on Android and HiDPI monitors as well.