CommandLineParser: pass options object as reference parameter [refactor]
Change the `ParseCommandLineOptions` method so that it modifies a
`CommandLineOptions` object passed in by reference, rather than
returning a new `CommandLineOptions` object after all command line
parsing has completed.

This change allows GameManager to access the intermediate state of the
command line options as they are being parsed, so that we can
implement order-dependent and interdependent options.

In particular, I made this change because I want to add a
`GltfImportTask.ProfilingEnabled` flag to control the profiling of
individual import tasks, but to do that unless I have access to the
intermediate state of the `--profile` flag during command line
parsing.

Note: This commit changes the behaviour of the `--profile` option from
a global, order-independent flag to an order-dependent flag that must
be appear *before* the `--import` or `--import-streaming-asset` that
should be profiled. Profiling can be disabled later in the command
line by using `--profile-`, as per the NDesk.Options documentation
[1]. (`--profile+` also works for enabling profiling.)

[1]: http://www.ndesk.org/doc/ndesk-options/NDesk.Options/OptionSet.html