CommandLineOptions: add `--log-message` option [new]
Add a `--log-message` option so that I can remove some hacky code in
GameManager.Update(), which logs a sentinel message `[PigletViewer]
IDLE` when all command-line tasks have completed.

I use the sentinel message when I'm profiling Piglet, in order to know
when it is safe to extract the profiling results from the Unity
log. But it is much cleaner to log the sentinel message using a
command-line option than to hard-code it in the GameManager Update
method.